Move vesselFieldValidators to separate folder to avoid merge conflicts and long file
