+++ src/lib/field-validators/vesselfieldValidators.ts
@@ -1,0 +1,48 @@
+import { FieldValidator } from "../fieldValidators";
+
+export class VesselMaxCapacityValidator extends FieldValidator {
+  constructor() {
+    super();
+    this._rules = [
+      {
+        errorMessage: "Maximum number of persons onboard is a required field",
+        predicateFn: (value) => value.length === 0,
+      },
+      {
+        errorMessage:
+          "Maximum number of persons onboard must be a whole number",
+        predicateFn: (value) => value.match(/\D+/) !== null,
+      },
+    ];
+  }
+}
+
+export class VesselAreaOfOperationValidator extends FieldValidator {
+  constructor() {
+    super();
+    this._rules = [
+      {
+        errorMessage: "Typical area of operation has too many characters",
+        predicateFn: (value) => value.length > 250,
+      },
+    ];
+  }
+}
+
+export class VesselBeaconLocationValidator extends FieldValidator {
+  constructor() {
+    super();
+    this._rules = [
+      {
+        errorMessage: "Where the beacon is kept has too many characters",
+        predicateFn: (value) => value.length > 100,
+      },
+    ];
+  }
+}
+
+export const fieldValidatorLookup = {
+  maxCapacity: new VesselMaxCapacityValidator(),
+  areaOfOperation: new VesselAreaOfOperationValidator(),
+  beaconLocation: new VesselBeaconLocationValidator(),
+};
